rename to filtered device list
