+++ res/css/components/views/settings/devices/_FilteredDeviceList.pcss
@@ -14,10 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-.mx_SortedDeviceList {
+.mx_FilteredDeviceList {
     list-style-type: none;
     display: grid;
     grid-gap: $spacing-16;
     margin: 0;
-    padding: 0;
+    padding: 0 $spacing-8;
 }
