+++ test/components/views/settings/devices/FilteredDeviceList-test.tsx
@@ -14,28 +14,18 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-.mx_DeviceTile {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
+import React from 'react';
+import { mount } from 'enzyme';
 
-    width: 100%;
-}
+import FilteredDeviceList from '../../../../../src/components/views/settings/devices/FilteredDeviceList';
 
-.mx_DeviceTile_info {
-    flex: 1 1 0;
-}
+describe('<FilteredDeviceList />', () => {
+    const defaultProps = {};
+    const getComponent = (props = {}) =>
+        mount(<FilteredDeviceList {...defaultProps} {...props} />);
 
-.mx_DeviceTile_metadata {
-    margin-top: 2px;
-    font-size: $font-12px;
-    color: $secondary-content;
-}
-
-.mx_DeviceTile_actions {
-    display: grid;
-    grid-gap: $spacing-8;
-    grid-auto-flow: column;
-
-    margin-left: $spacing-8;
-}
+    it('renders', () => {
+        const component = getComponent();
+        expect(component).toBeTruthy();
+    });
+});
